third_party/blink/web_tests/external/wpt/editing/run/unlink-expected.txt

This is a testharness.js-based test.
Found 23 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["unlink",""]] "foo[]bar": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<b>foo[]bar</b>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<i>foo[]bar</i>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<span>foo</span>{}<span>bar</span>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<span>foo[</span><span>]bar</span>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<a href=http://www.google.com/>foo[bar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "<a href=\\"http://www.google.com/\\">foo</a>bar<a href=\\"http://www.google.com/\\">baz</a>"
[FAIL] [["unlink",""]] "<a href=http://www.google.com/>foo[barbaz</a>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "<a href=\\"http://www.google.com/\\">foo</a>barbaz"
[FAIL] [["unlink",""]] "{<a href=http://www.google.com/>foobar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foobar<a href=\\"http://www.google.com/\\">baz</a>"
[FAIL] [["unlink",""]] "foo<a href=http://www.google.com/>b[]ar</a>baz": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "foo<a href=http://www.google.com/>b[]ar</a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<a href=\\"http://www.google.com/\\">bar</a>baz"
[FAIL] [["unlink",""]] "<a id=foo href=http://www.google.com/>foobar[]baz</a>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<a id=foo href=http://www.google.com/>foobar[]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a id=\\"foo\\">foobarbaz</a>" but got "<a id=\\"foo\\" href=\\"http://www.google.com/\\">foobarbaz</a>"
[FAIL] [["unlink",""]] "<a id=foo href=http://www.google.com/>foo[bar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a id=\\"foo\\">foobarbaz</a>" but got "<a id=\\"foo\\" href=\\"http://www.google.com/\\">foo</a>bar<a href=\\"http://www.google.com/\\">baz</a>"
[FAIL] [["unlink",""]] "<a id=foo href=http://www.google.com/>[foobarbaz]</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a id=\\"foo\\">foobarbaz</a>" but got "foobarbaz"
[FAIL] [["unlink",""]] "foo<a id=foo href=http://www.google.com/>[bar]</a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a id=\\"foo\\">bar</a>baz" but got "foobarbaz"
[FAIL] [["unlink",""]] "foo[<a id=foo href=http://www.google.com/>bar</a>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a id=\\"foo\\">bar</a>baz" but got "foobarbaz"
[FAIL] [["unlink",""]] "[foo<a id=foo href=http://www.google.com/>bar</a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a id=\\"foo\\">bar</a>baz" but got "foobarbaz"
[FAIL] [["unlink",""]] "<a name=foo>foobar[]baz</a>": execCommand("unlink", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["unlink",""]] "<a name=foo>foo[bar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a name=\\"foo\\">foobarbaz</a>" but got "<a name=\\"foo\\">foo</a>bar<a name=\\"foo\\">baz</a>"
[FAIL] [["unlink",""]] "<a name=foo>[foobarbaz]</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a name=\\"foo\\">foobarbaz</a>" but got "foobarbaz"
[FAIL] [["unlink",""]] "foo<a name=foo>[bar]</a>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a name=\\"foo\\">bar</a>baz" but got "foobarbaz"
[FAIL] [["unlink",""]] "foo[<a name=foo>bar</a>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a name=\\"foo\\">bar</a>baz" but got "foobarbaz"
[FAIL] [["unlink",""]] "[foo<a name=foo>bar</a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a name=\\"foo\\">bar</a>baz" but got "foobarbaz"
Harness: the test ran to completion.
